+++ toxcore/Messenger.h
@@ -208,7 +208,7 @@ typedef struct Messenger {
     void *friend_statusmessagechange_userdata;
     void (*friend_userstatuschange)(struct Messenger *m, int, USERSTATUS, void *);
     void *friend_userstatuschange_userdata;
-    void (*friend_typingchange)(struct Messenger *m, int, uint8_t, void *);
+    void (*friend_typingchange)(struct Messenger *m, int, int, void *);
     void *friend_typingchange_userdata;
     void (*read_receipt)(struct Messenger *m, int, uint32_t, void *);
     void *read_receipt_userdata;
@@ -461,9 +461,9 @@ void m_callback_statusmessage(Messenger *m, void (*function)(Messenger *m, int, 
 void m_callback_userstatus(Messenger *m, void (*function)(Messenger *m, int, USERSTATUS, void *), void *userdata);
 
 /* Set the callback for typing changes.
- *  Function(int friendnumber, uint8_t is_typing)
+ *  Function(int friendnumber, int is_typing)
  */
-void m_callback_typingchange(Messenger *m, void(*function)(Messenger *m, int, uint8_t, void *), void *userdata);
+void m_callback_typingchange(Messenger *m, void(*function)(Messenger *m, int, int, void *), void *userdata);
 
 /* Set the callback for read receipts.
  *  Function(int friendnumber, uint32_t receipt)
