+++ toxcore/tox.h
@@ -352,9 +352,9 @@ void tox_callback_status_message(Tox *tox, void (*function)(Tox *tox, int, uint8
 void tox_callback_user_status(Tox *tox, void (*function)(Tox *tox, int, TOX_USERSTATUS, void *), void *userdata);
     
 /* Set the callback for typing changes.
- *  function (int friendnumber, uint8_t is_typing)
+ *  function (int friendnumber, int is_typing)
  */
-void tox_callback_typing_change(Tox *tox, void (*function)(Tox *tox, int, uint8_t, void *), void *userdata);
+void tox_callback_typing_change(Tox *tox, void (*function)(Tox *tox, int, int, void *), void *userdata);
 
 /* Set the callback for read receipts.
  *  function(int friendnumber, uint32_t receipt)
